Fix: Handle large integers for alarm order
Change the `order` field in the `Alarm` model from `i32` to `i64`. This prevents deserialization errors when the frontend sends `Date.now()` (milliseconds since epoch) as the `order` value for new alarms, which overflows a 32-bit signed integer.

+++ src-tauri/src/models.rs
@@ -7,7 +7,7 @@ pub struct Alarm {
     pub repeat_type: RepeatType,
     pub triggers: Vec<TriggerInfo>,
     pub enabled: bool,
-    pub order: i32,
+    pub order: i64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
